주석 추가

+++ app/src/main/java/com/example/movie_app/data/value_object/MovieDetails.kt
@@ -1,22 +1,22 @@
-package com.example.movie_app.data.value_object
-import com.google.gson.annotations.SerializedName
+    package com.example.movie_app.data.value_object
+    import com.google.gson.annotations.SerializedName
 
-// 유저가 선택한 영화에 상세 정보를 나타낼 때 필요한 데이터
-data class MovieDetails(
-    val budget: Int,
-    val id: Int,
-    val overview: String,
-    val popularity: Double,
-    @SerializedName("poster_path")
-    val posterPath: String,
-    @SerializedName("release_date")
-    val releaseDate: String,
-    val revenue: Int,
-    val runtime: Int,
-    val status: String,
-    val tagline: String,
-    val title: String,
-    val video: Boolean,
-    @SerializedName("vote_average")
-    val rating: Double
-)
+    // 유저가 선택한 영화에 상세 정보를 나타낼 때 필요한 데이터
+    data class MovieDetails(
+        val budget: Int,
+        val id: Int,
+        val overview: String,
+        val popularity: Double,
+        @SerializedName("poster_path")
+        val posterPath: String,
+        @SerializedName("release_date")
+        val releaseDate: String,
+        val revenue: Long,
+        val runtime: Int,
+        val status: String,
+        val tagline: String,
+        val title: String,
+        val video: Boolean,
+        @SerializedName("vote_average")
+        val rating: Double
+    )
